refactor(root): Move Combine extensions to a new file
